requirements.txt
python-docx==1.1.0
pytz==2022.7
pandas==1.5.3
streamlit==1.33.0
